+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 ### Added
+- Added `sync` command to stage markdown from a source directory with link and attachment rewriting ([#1](https://github.com/kcosr/confluence-md/pull/1))
 
 ### Changed